feat(escpos): add initial ESC/POS command implementations
- Introduced various command files for ESC/POS functionality

This lays the groundwork for further development of the ESC/POS protocol
