+++ pkg/posprinter/protocol/escpos/alerts.go
@@ -1,0 +1,6 @@
+package escpos
+
+// TODO: Comandos para generar alertas audibles
+// - Pitidos
+// - Patrones de sonido
+// - Alertas de error
