+++ pkg/posprinter/protocol/escpos/cash_drawer.go
@@ -1,0 +1,10 @@
+package escpos
+
+// TODO: Comandos para control del cajón de efectivo
+// - Apertura de cajón
+// - Estado de cajón
+
+func (p *Commands) Pulse(pin int, onMS int, offMS int) []byte {
+	// TODO: Implementar ESC p m t1 t2
+	return []byte{}
+}
