+++ pkg/posprinter/protocol/escpos/doc.go
@@ -3,10 +3,10 @@ Package escpos proporciona herramientas para interactuar con impresoras térmica
 
 Este paquete ofrece:
 
-  - Implementación del protocolo ESC/POS para la generación de comandos.
-  - Funciones para inicializar, configurar y gestionar la impresión de texto e imágenes.
-  - Métodos para manejar comandos de corte, alimentación, y formateo de datos.
-  - Utilidades para procesar imágenes, convirtiéndolas en datos raster compatibles.
-  - Conjunto de funciones de bajo nivel para emitir secuencias de comandos y consultar el estado de la impresora.
+- Implementación del protocolo ESC/POS para la generación de comandos.
+- Funciones para inicializar, configurar y gestionar la impresión de texto e imágenes.
+- Métodos para manejar comandos de corte, alimentación, y formateo de datos.
+- Utilidades para procesar imágenes, convirtiéndolas en datos raster compatibles.
+- Conjunto de funciones de bajo nivel para emitir secuencias de comandos y consultar el estado de la impresora.
 */
 package escpos
