+++ pkg/posprinter/protocol/escpos/hardware_settings.go
@@ -1,0 +1,7 @@
+package escpos
+
+// TODO: Comandos para configuración de hardware específico
+// - Densidad de impresión
+// - Velocidad de impresión
+// - Modo de ahorro de energía
+// - Control de periféricos
