+++ pkg/posprinter/protocol/escpos/memory.go
@@ -1,0 +1,7 @@
+package escpos
+
+// TODO: Comandos para manejo de memoria interna de la impresora
+// - Definición de caracteres personalizados
+// - Almacenamiento de logotipos
+// - Carga de imágenes en memoria
+// - Selección de memoria activa
