+++ pkg/posprinter/protocol/escpos/status.go
@@ -1,0 +1,8 @@
+package escpos
+
+// TODO: Comandos para obtener estado de la impresora
+// - Consulta de estado en tiempo real
+// - Sensores de papel
+// - Estado de error
+// - Transmisión de estado
+// - Autodiagnóstico
