+++ pkg/posprinter/protocol/escpos/vendor_specific.go
@@ -1,0 +1,7 @@
+package escpos
+
+// TODO: Comandos propietarios específicos de fabricantes
+// - Extensiones Epson
+// - Extensiones Star
+// - Extensiones Bixolon
+// - Otros fabricantes
